fix: separate riff loader
